% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Marwick-WII13-ochre.R
\name{calibration}
\alias{calibration}
\title{Calibration}
\usage{
calibration(combys9, remaining)
}
\description{
Calibration
}
\examples{
\dontrun{
 combys12 <- calibration(combys9$combys9, combys9$remaining)
}
}

