% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marwick-WII13-ochre.R
\name{channels_and_AUC}
\alias{channels_and_AUC}
\title{Get pXRF channels and compute AUC for those elements}
\usage{
channels_and_AUC(df, files)
}
\description{
Get pXRF channels and compute AUC for those elements
}
\examples{
\dontrun{
aucAllSamples <- channels_and_AUC(df$df, df$files)
}
}

