% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Marwick-WII13-ochre.R
\name{corr_with_Fe}
\alias{corr_with_Fe}
\title{Compute correlations with Fe}
\usage{
corr_with_Fe(combys9)
}
\description{
Compute correlations with Fe
}
\examples{
\dontrun{
combys9 <- corr_with_Fe(combys9)
}
}

