% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Marwick-WII13-ochre.R
\name{mag_sus_corr}
\alias{mag_sus_corr}
\title{Mag sus corr}
\usage{
mag_sus_corr(ochre_data)
}
\description{
Mag sus corr
}
\examples{
\dontrun{
HDI_corr <- mag_sus_corr(ochre_data)
}
}

