% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Marwick-WII13-ochre.R
\name{ochre_table}
\alias{ochre_table}
\title{Table summarising ochre data}
\usage{
ochre_table(ochre_data, pXRF_data)
}
\description{
Table summarising ochre data
}
\examples{
\dontrun{
the_ochre_table <- ochre_table(ochre_data, combys12$combys11)
}
}

