% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marwick-WII13-ochre.R
\name{subset_5_runs}
\alias{subset_5_runs}
\title{Subset 5 runs from the 10 from each specimen}
\usage{
subset_5_runs(aucAllSamples, elementList)
}
\description{
Subset 5 runs from the 10 from each specimen
}
\examples{
\dontrun{
 combys9 <- subset_5_runs(aucAllSamples$aucAllSamples, aucAllSamples$elementList)
 }
}

